backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
psycopg2-binary>=2.9.7
pydantic>=2.4.2
PyJWT>=2.8.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
anthropic>=0.52.2
groq>=0.9.0
python-multipart==0.0.9
httpx==0.26.0
